[REFACTOR] 마이페이지 조합 카드 적용 (#117)
